# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Campaign Service",
    "description": "API for managing marketing campaigns",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3003"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service API key (CAMPAIGN_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "StatsResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/StatsEntry"
          }
        },
        "required": [
          "stats"
        ]
      },
      "StatsEntry": {
        "type": "object",
        "properties": {
          "totalCampaigns": {
            "type": "number"
          },
          "byStatus": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "budgetTotalUsd": {
            "type": "number",
            "nullable": true
          },
          "maxLeadsTotal": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "totalCampaigns",
          "byStatus",
          "budgetTotalUsd",
          "maxLeadsTotal"
        ]
      },
      "GroupedStatsResponse": {
        "type": "object",
        "properties": {
          "groupedStats": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/StatsEntry"
            }
          }
        },
        "required": [
          "groupedStats"
        ]
      },
      "StatsGroupByEnum": {
        "type": "string",
        "enum": [
          "workflowSlug",
          "featureSlug",
          "workflowDynastySlug",
          "featureDynastySlug"
        ]
      },
      "BatchBudgetUsageBody": {
        "type": "object",
        "properties": {
          "campaignIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1
          }
        },
        "required": [
          "campaignIds"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/stats": {
      "get": {
        "tags": [
          "Stats"
        ],
        "summary": "Campaign stats from own DB (query params)",
        "description": "Returns campaign counts, status breakdown, and configured budget totals. Supports filtering by workflowSlug, featureSlug, workflowDynastySlug, featureDynastySlug, and groupBy for aggregation by slug or dynasty slug. When groupBy is set, returns groupedStats keyed by the group value. Requires API key.",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "orgId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "featureSlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "featureDynastySlug",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/StatsGroupByEnum"
            },
            "required": false,
            "name": "groupBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign stats (flat or grouped)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/StatsResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GroupedStatsResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stats/batch-budget": {
      "post": {
        "tags": [
          "Stats"
        ],
        "summary": "Get cost data for multiple campaigns",
        "description": "Returns budget usage (totalCostInUsdCents) per campaign via runs-service.",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchBudgetUsageBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stats per campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "maxLeads": {
                            "type": "number",
                            "nullable": true
                          },
                          "maxBudgetTotalUsd": {
                            "type": "string",
                            "nullable": true
                          },
                          "totalCostInUsdCents": {
                            "type": "string",
                            "nullable": true
                          },
                          "error": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "results"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}